requirements.txt
# Optimized for Render hosting with Firebase
discord.py==2.3.2
python-dotenv==1.0.0
aiohttp==3.9.1
firebase-admin==6.2.0 